Move City bases into cities/base_cities.py
